feat: Add 'collection/:collectionAddress' route and view (#342)
