+++ src/components/Collections/Catalog/index.tsx
@@ -1,37 +1,23 @@
 import React, { useEffect } from 'react';
-import { Box, Flex, Heading, Text, Link, Spinner } from '@chakra-ui/react';
+import { Flex } from '@chakra-ui/react';
 import { useLocation } from 'wouter';
-import { RefreshCw, ExternalLink } from 'react-feather';
-import { MinterButton } from '../../common';
 import Sidebar from './Sidebar';
-import TokenGrid from './TokenGrid';
-import CollectionsDropdown from './CollectionsDropdown';
-
+import CollectionDisplay from './CollectionDisplay';
 import { useSelector, useDispatch } from '../../../reducer';
-import {
-  getContractNftsQuery,
-  getWalletAssetContractsQuery
-} from '../../../reducer/async/queries';
+import { getWalletAssetContractsQuery } from '../../../reducer/async/queries';
 import { selectCollection } from '../../../reducer/slices/collections';
 
 export default function Catalog() {
   const [, setLocation] = useLocation();
-  const { system, collections: state } = useSelector(s => s);
+  const system = useSelector(s => s.system);
+  const collections = useSelector(s => s.collections);
   const dispatch = useDispatch();
 
   useEffect(() => {
-    const selectedCollection = state.selectedCollection;
-    if (selectedCollection === null) {
-      dispatch(selectCollection(state.globalCollection));
-    } else {
-      dispatch(getContractNftsQuery(selectedCollection));
+    if (collections.selectedCollection === null) {
+      dispatch(selectCollection(collections.globalCollection));
     }
-  }, [
-    system.status,
-    state.selectedCollection,
-    state.globalCollection,
-    dispatch
-  ]);
+  }, [collections.selectedCollection, collections.globalCollection, dispatch]);
 
   useEffect(() => {
     if (system.status !== 'WalletConnected') {
@@ -41,12 +27,10 @@ export default function Catalog() {
     }
   }, [system.status, setLocation, dispatch]);
 
-  const selectedCollection = state.selectedCollection;
+  const selectedCollection = collections.selectedCollection;
   if (system.status !== 'WalletConnected' || !selectedCollection) {
     return null;
   }
-
-  const collection = state.collections[selectedCollection];
 
   return (
     <Flex
@@ -70,84 +54,7 @@ export default function Catalog() {
       >
         <Sidebar />
       </Flex>
-      <Flex
-        flexDir="column"
-        h="100%"
-        w="100%"
-        px={{ base: 6, md: 10 }}
-        pt={6}
-        flex="1"
-        bg="brand.brightGray"
-        borderLeftWidth="1px"
-        borderLeftColor="brand.lightBlue"
-        overflowY="scroll"
-        justify="start"
-      >
-        <Flex display={{ base: 'flex', md: 'none' }} mb={4}>
-          <CollectionsDropdown />
-        </Flex>
-        <Flex
-          display={{ base: 'none', md: 'flex' }}
-          w="100%"
-          pb={6}
-          justify="space-between"
-          align={{
-            base: 'flex-start',
-            md: 'center'
-          }}
-          flexDir={{
-            base: 'column',
-            md: 'row'
-          }}
-        >
-          <Flex flexDir="column" width="100%">
-            <Flex justify="space-between" width="100%">
-              <Heading size="lg">{collection.metadata.name || ''}</Heading>
-            </Flex>
-            <Flex align="center">
-              <Text fontFamily="mono" color="brand.lightGray">
-                {collection.address}
-              </Text>
-              <Link
-                href={system.config.bcd.gui + '/' + collection.address}
-                color="brand.darkGray"
-                isExternal
-                ml={2}
-              >
-                <ExternalLink size={16} />
-              </Link>
-            </Flex>
-          </Flex>
-          <MinterButton
-            variant="primaryActionInverted"
-            onClick={() => {
-              const selectedCollection = state.selectedCollection;
-              if (selectedCollection !== null) {
-                dispatch(getContractNftsQuery(selectedCollection));
-              }
-            }}
-            mt={{
-              base: 4,
-              md: 0
-            }}
-          >
-            <Box color="currentcolor">
-              <RefreshCw size={16} strokeWidth="3" />
-            </Box>
-            <Text ml={2}>Refresh</Text>
-          </MinterButton>
-        </Flex>
-        {!collection.loaded ? (
-          <Flex flexDir="column" align="center" flex="1" pt={20}>
-            <Spinner size="xl" mb={6} color="gray.300" />
-            <Heading size="lg" textAlign="center" color="gray.500">
-              Loading...
-            </Heading>
-          </Flex>
-        ) : (
-          <TokenGrid state={state} walletAddress={system.tzPublicKey} />
-        )}
-      </Flex>
+      <CollectionDisplay address={selectedCollection} />
     </Flex>
   );
 }
